requirements.txt
discord.py
python-dotenv
flask
pymongo
dnspython
certifi
